Scope epics internal_ids generation to `issues` usage
We work on simlessly migrating Epics to WorkItems. Because we
currently have no work items available at group level we will
actually migrate Epics as the first group level WorkItem and as
an added bonus we get to keep Epic's IIDs for the group level
Epic WorkItem IID.

Because links to epics use IID, this would allow us to maintain
compatibility between old and new links with a simple redirect as
the IID we need to use to lookup the record would be the same.

In order to achieve that and unblock creation of other group level
WorkItems, we want to make sure that IID allocation for Epics is now
using the `issues` usage in `inteternal_ids` table.

In order for Epic model to allocate IID based on `issues` usage, we
need to override the `internal_id_scope_usage` method to return `issues`.
This however creates a challenge in rolling deployment environments
where some application nodes can run on "old code" while other
application nodes would run on "new code", thus creating a situation
where Epic model with "old code" would try to allocate IID by
looking up a `usage: :epics` record in `internal_ids` while
"new code" would lookup `usage: :issues` record in `internal_ids`
which can lead to duplicate IID allocation and application errors.

To solve that we are adding a couple triggers to run on insert/update
to keep the `last_value` column for the records for the two usages:
`issues` and `epics` in sync.

These triggers are only needed for the rolling deployment duration
when we can have 2 versions of the code running.

Once the new code is deployed Epic model will always allocate IID by
using the issues usage and triggers can be removed.

Changelog: changed