Fetch tag notes on release refresh
We want to ensure that when a draft release is loaded from local
storage, that we fetch the tag details associated with the pre-loaded
tag. This ensures that the tag message is loaded and the "existing tag
in use" message is displayed if another release is created in the
mean-time.

Changelog: fixed