Add only_new_paths option to Tooling::FindChanges
When a file in the documentation is moved, the old
path was also passed to markdownlint and vale.
That causes the linting to fail, because the files
don't exist.

In order to fix that, the new option will only return
the new paths of the changed files and not the old paths.