Add a constraint to ensure application_settings.rate_limits is a hash
In f89cc1649a0559f0eaaefbdd9792be8558f67134, we introduced a JSONB
column `rate_limits` to store the application rate limits. We want to
ensure that we always store a hash in this column otherwise, the rails
application will break. This can happen if a bad update query is
executed like in the following case if we redefine the
application_settings class:

```ruby
class MyMigration

  class TmpSetting < ApplicationRecord
    self.table_name = :application_settings
  end

  def up
    TmpSetting.update(rate_limits: [1])
  end
end
```

Changelog: other