feat: Add QA to release-environments pipeline
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2825

In order to actually have release-environments be tested and validate
that the merged code on the stable branch is functional, we want
to add a step to the release-environments child pipeline to run
GitLab QA against the release-environment. This will happen after
deploy.

As a first iteration, we want the test to just be smoke tests,
as we expect some failures and further iteration to ensure the
release-environments are ready for bigger test suites.