Force diff output to color gitlab-housekeeper preview
The diff output does not use color when shelled out from Ruby by
default. Setting `--color=always` seems to fix this. Since this is only
used to send directly to `puts` straight after this also seems safe. If
we were using this diff content for something else it might not be
great but the scope of this line of code is pretty clearly only for the
preview output.