Do not restrict on CI_PROJECT_PATH for Gitaly test
CI_PROJECT_PATH should be set as the path where it's running the
pipeline, and in the case of Gitaly triggering a cross project
pipeline, it is `gitlab-org/gitlab` rather than `gitlab-org/gitaly`.

That said, checking the variable doesn't really make sense, because
it's likely the pipeline wouldn't really run if anything else than
`gitlab-org/gitlab` is set. This variable is used in a lot of places,
changing it will likely to break random things, so it should be just
assumed and it should not be changed.

Also CI_PIPELINE_SOURCE should be `pipeline` for cross project
pipelines.