Run most FOSS jobs in FOSS pipeline
Note that this also removes clusterwide-db for FOSS because it's not
passing for EE anyway.

Also note that this can make the pipelines run much more FOSS jobs
because previously we might only run partial FOSS jobs like
build-assets-image, but now when we run it will run all the
corresponding FOSS jobs.

We can tweak the rules accordingly in the next iteration. For this
iteration this is focusing on bringing everything over, and it should
avoid missing jobs unless there's a good reason like clusterwide-db jobs
can't pass anyway.