Make issuable state declarations consistent
For Issue and MergeRequest, we declare state methods in Issuable that
work like that of ActiveRecord enums. We also depend on state_machine to
declare methods like opened? and closed?

For Epic, we actually declare an enum on state_id and then alias it to
state.

This makes things consistent by changing Epic to use the same approach
used by Issue and MergeRequest.