Add standard description/labels to MRs and purple color titles
This adds the `gitlab-housekeeper-authored` label to all MRs and ensures
the commit message and MR always have the housekeeper link included so
that we don't need to add that manually in the keeps anymore.

It also adds some `purple` color to the dry-run output for the titles. This
makes them easier to read.