Remove true-up validation for cloud licenses
Prior to this change, when a cloud license subscription was renewed with
an insufficient true-up count, the renewal license (either by applying a
new offline license key or by activating/syncing an online cloud
subscription) couldn't be applied on the SM instance.

This change removes the true-up validation for cloud licenses in order
to reduce disruption after a renewal.

Changelog: changed
EE: true