Update nullable on author field on NoteType
We cannot guarantee that note author will not be null as we want to
keep track of the system changes happening on work items. So when a user
is completelly removed from the system, i.e. removed from DB we still
want to keep historical information about what happened on a specific
work item so we would show a "delete user" info.

https://gitlab.com/gitlab-org/gitlab/-/issues/438713

Changelog: fixed