Remove `@include`s from SCSS files
This replaces a bunch of `@include` rules with their actual CSS
definitions. This is done in preparation for moving to Tailwind CSS,
which won't let us include utilities the same way due to our inability
to support the `@apply` directive.

The `@include` approach was already questioned in the past as it doen't
have much value over explicit style definitions and just adds an
unnecessary abstraction layer.