Remove @service instance var
This is forward refactoring for
https://gitlab.com/gitlab-org/gitlab/-/issues/391526 where the
GitLab for Slack app integration can be associated with a group also.

We won't add a `has_one :gitlab_slack_application_integration`
association to `Group` as this is discouraged and we can avoid needing
to do so if we instead look up the integration through
the `SlackIntegration` record.