Add buffer_size to SDK client arguments for performance
This adds `buffer_size` of 100 to Snowplow events being sent by the
Internal Events framework to Product Analytics for performance reasons.

Currently, every event is sent in its own request, but with this
argument passed, the events will be batched and sent in a single POST
request only after there are enough events in the buffer.