Improves documentated description of BranchRuleType
The description for BranchRuleType was describing a list but the
description should be for each object.

Also clarifies the name field's description as this field is used not
just for specific branch names or wildcard names, it is also used for
"All branches" and "All protected branches" so we've decided to relabel
the value in the UI "Rule Target". The new description better reflects
this line of thinking.