Define global identifiers for branch rules
We are implementing graphql mutations to update and delete branch rules.
To do this we need identifiers that we can pass to the mutations. This
change uses the ProtectedBranch ID to identify Projects::BranchRule
objects and Project ID to identify Projects::AllBranchesRule and
Projects::AllProtectedBranchesRule objects.

This change also adds a `::find(id)` method to the BranchRule classes
allowing us to use global ids to find/intialize the objects.

The Global IDs will take the following formats
- `gid://gitlab/Projects::BranchRule/#{protected_branch.id}`
- `gid://gitlab/Projects::AllBranchesRule/#{project.id}`
- `gid://gitlab/Projects::AllProtectedBranchesRule/#{project.id}`

Changelog: added