Migrates index_ml_models to graphql
To avoid technical debt and improve user experience, this MR migrates
index_ml_model.vue to use graphql instead of preloading and refreshing
the page on every change to the search query and pagination.

To do so, we introduce search to the SearchableList component, in a way
that doesn't affect existent usage.