Add `include_archived` param to the `Namespaces::ProjectsFinder`
Context
----------

For the group-level view of the Vulnerability Report, we exclude
archived projects.

However, the project filter list includes archived projects as
selections

We want to exclude archived projects from that list.

The list is populated by a graphql
call ([`security_dashboard/graphql/queries/group_projects.query.graphql`][1])

Which ultimately ends up calling this finder in the
[`namespace_projects_resolver`][0].

This Commit
----------

This commit adds a param to the finder to allow us to eventually filter
out archived projects in our graphql query.

As the existing behavior of this finder is to include archived
projects, the default for the new param is `include_archived: true` to
keep any other usages of this finder unaffected.

[0]:https://gitlab.com/gitlab-org/gitlab/-/blob/b1f2a2d85e2f920b1e0a98f71d1d29a9aed3fc5c/app/graphql/resolvers/namespace_projects_resolver.rb
[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/b1f2a2d85e2f920b1e0a98f71d1d29a9aed3fc5c/ee/app/assets/javascripts/security_dashboard/graphql/queries/group_projects.query.graphql#L4

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/426520