Add `includeArchived: true | false` to `groupProjects` graphQL
This Commit
-----------------------

This commit exposes the new `include_archived` filter param in the
`groupProjects` graphql resolver that was added in [MR 141989][2]

As the existing behavior was to include archived projects, the default
for the new param is `include_archived: true` to keep any other
usages of this graphql resolver unaffected.

Context
-----------------------

For the group-level view of the Vulnerability Report, we exclude
archived projects.

However, the project filter list includes archived projects as
selections:

![image](/uploads/451fd85e4848f1e7ad361eaee4299aad/image.png)

We want to exclude archived projects from that list.

The list is populated by a graphql
resolver ([`security_dashboard/graphql/queries/group_projects.query.graphql`][1])

That resolver ultimately ends up calling the finder which was updated
in [MR 141989][2] to accept a new filter param (`include_archived`)
([`src`][0]).

[0]:https://gitlab.com/gitlab-org/gitlab/-/blob/b1f2a2d85e2f920b1e0a98f71d1d29a9aed3fc5c/app/graphql/resolvers/namespace_projects_resolver.rb
[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/b1f2a2d85e2f920b1e0a98f71d1d29a9aed3fc5c/ee/app/assets/javascripts/security_dashboard/graphql/queries/group_projects.query.graphql#L4
[2]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/141989

resolves: https://gitlab.com/gitlab-org/gitlab/-/work_items/438814
related to: https://gitlab.com/gitlab-org/gitlab/-/issues/437004

Changelog: changed