Remove Frontend Implementation of sast_reports_in_inline_diff Flag
This MR completely removes the frontend implementation of the sast_reports_in_inline_diff feature flag.
It includes the elimination of all conditional logic that was used to toggle between the old and new versions of inline findings.
Furthermore, this change also encompasses the removal of the frontend code related to the older inline-findings implementation.

EE: true