Validate Direct Transfer GraphQL query deprecation
This adds a `shared_examples` to validate that Direct Transfer GraphQL
queries do not use deprecated fields or arguments.

It piggy-backs on the existing analyzer that adds this data to the
`RequestStore`. It validates the assumptions that certain keys are
present in this store.

The intention of the spec is for future breaking changes to GraphQL
schema used by Direct Transfer queries to become known as soon as
possible. Engineers can then work out how best to proceed with the
deprecation.

https://gitlab.com/gitlab-org/gitlab/-/issues/438021