Refactor Gitlab::License factory
Prior to this change, the factories for `Gitlab::License` and `License`
where defined in the same file (`ee/spec/factories/licenses.rb`).

This change moves the factory for `Gitlab::License` into its own file.
Additionally, the default value for the plan is updated from the
deprecated Starter plan to a Premium plan.