Fix error in Mapper::Normalizer when CI include value type is invalid
When the CI 'include' value is not a string nor a hash
it raises an 'undefined method 'deep_symbolize_keys'' error
in Ci::Config::External::Mapper::Normalizer. This MR fixes
this by adding value type validation to the class. It now
returns a validation error instead of failing.

Changelog: fixed