Use static list for dependency list licenses
The current implementation of the licenses list gets all of the unique
licenses in the project, which potentially means scanning hundreds of
thousands of rows in order to find the unique licenses, which is very
slow. This change improves the performance by load the licenses from a
static list instead.

Changelog: fixed
EE: true