Allow pajamas_adoption job to block merge requests
This updates the `pajamas_adoption` job to allow it to block merge
requests if they regress the adoption of the Pajamas Design System.

Unexpected errors cause it to exit with status 1, which will not block
merge requests. An example of this would be when the merge request's
source branch has more than 20 commits. This causes the scan to fail,
but should not block the merge request.

Part of https://gitlab.com/gitlab-com/gitlab-OKRs/-/work_items/5345.