Update BulkImports::Failure to display subrelation that failed to import
Direct Transfer API shows failures for the things that we could not
import. If a subrelation of a top level relation failed to import
(e.g. MR diff note) - we still show top level relation as the one
that failed to import. In reality, the top level relation is imported,
but a subrelation did not.

Update relation to include the failed to import subrelation, so it's
more obvious what exactly failed to import.

Changelog: added