Fully deprecate `public_builds` in projects API
When we changed from `public_builds` to `public_jobs` in https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/9463
we missed the `PUT` and `POST` endpoints. Internally we still refer to
it as `public_builds` and due to the breaking changes policy for the API
we will still need to support `public_builds` as an API parameter, so
using the `translate_params_for_compatibility()` method to support both
is a convenient way to achieve consistency without a large refactor.

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/329725