Remove transaction around object pool mirroring
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/439381

**Problem**

Replication of the object pool might take a significant time for large
repositories. If the waiting time exceed 30 seconds, then the Postgres
transaction will be terminated and lead to a migration failure.

**Solution**

We can avoid a transaction usage, but then we should manually remove a
newly created `PoolRepository` in case of the error.

Changelog: fixed