Use group ID instead of root group ID for query
In the backend counterpart MR, it has been decided to use the group ID
instead of the top-level group ID for getting the SAML users:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/141040#note_1743246071

This API query is not yet been used because it is under a feature flag,
`group_users_saml`, that is currently off in all environments. So it is
safe to deploy.