Fix Members::DestroyService audit bug with removing subresources
Fixes a bug where the `Members::DestroyService` would be called
recursively with no `current_user` and attempt to log an audit event
without an `author` parameter. This was encountered by a customer when
running the `RemoveExpiredMembersWorker` .

Also adds full error tracking for the blind rescue in
`RemoveExpiredMembersWorker` . The blind rescue and limited logging made
it difficult to understand the customer circumstances and reproduce the
error.

Changelog: fixed
EE: true