Update links in metadata only + trailing spaces
Drop the scope of this merge request to be nothing but changing
the URLs in the metadata to be correct, and then fixing any trailing
space problems so lefthook will pass.

That makes this first MR easier to review, and the remaining hits
are ones that we ACTUALLY need to review by hand.