Change purchase logic for product analytics
Previously we ignored the first purchased add-on and
calculated storage limits by multiplying (N-1)*1000000.

We've updated CDot so that it passes through the
correct number so we no longer need to
change the logic in the rails monolith

No customers affected by this, the feature is beta
and billing is behind a default-off feature flag.