Preload downstream_pipeline in StageEntity
This fixes an N+1 for the commit status dropdown in places such as the
stages controller, which is used to display a list of jobs on pipelines.

The additional preloads are needed for policy checks and path rendering.
In order to make this work, we also had to extend `CommitStatusPreloader`
with support for nested associations.

The change is behind the feature flag
`preload_ci_bridge_downstream_pipelines`, disabled by default.