Remove unused endpoint URL
The property
`Vulnerabilities::FeedbackEntity#destroy_vulnerability_feedback_dismissal_path`
is no longer referenced in the code base and can be removed.
The last code referencing this property was removed by
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/125362

The motivation for this is the removal of all uses of route helpers
referencing the deprecated controller
`Projects::VulnerabilityFeedbackController`.
This is a prerequisite for the removal of the controller itself as
described in https://gitlab.com/gitlab-org/gitlab/-/issues/419214