Add --filter-identifiers option to gitlab-housekeeper
During local usage of this we found it is often
[desirable to get the housekeeper to generate a specific MR](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/139747#note_1741757815)
and ignore the other ones. Since it just does them in order we had to
previously manually hack the keeps to skip the changes we don't care
about.

This change introduces a command line option `--filter-identifiers`
which takes a comma separated list which will be matched against the
list of "identifiers" in a change yielded from a keep. The identifiers
is an array so to allow for flexible regex matching of specific elements
in the array we filter such that all regexes must match at least one of
the identifiers in the list.