Remove --require option from gitlab-housekeeper but load all keeps/*.rb
This `-r keeps/...` is proving to be unnecessary and likely it's fine to
just always load all the files in `keeps/*.rb`.

I considered the possibility of loading everything in `keeps/*.rb` as
well as keeping the option around in case someone wanted to manually
load a keep that was in a different directory but I couldn't think of
any good reason to do that. So to simplify the options I've just removed
it.