Allow new guest users exceeding instance user cap on ultimate license
This is behind a feature flag.
When the feature flag is enabled, new non-billable users will be set
to active even if they exceed the instance user cap.
With the feature flag disabled, any new human user, billable or not,
will be blocked pending approval if the instance user cap is exceeded.