Remove archived projects from the group vulnerability report filter
This commit uses the `includeArchived: false` filter to exclude archived
projects from the filter list on the group-level vulnerabilities
dashboard

Context
------------------

We [do not include][0] archived projects in the group-level
vulnerabilities report (e.g. in the summary counts, in the list of
vulnerabilities).

However, archived projects show up in the project filter as an option
to filter by.

A new filter param, `includeArchived`, was added to the
`groupProjects` graphql resolver used by this filter in:

1. https://gitlab.com/gitlab-org/gitlab/-/merge_requests/142396
1. https://gitlab.com/gitlab-org/gitlab/-/merge_requests/141989

[0]:https://gitlab.com/gitlab-org/gitlab/-/issues/213144

resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/426520

Changelog: fixed
EE: true