Fix webhooks to POST correct MR state after create
Previously `AfterCreateService` was triggering hooks and integrations
early in the service. This meant the hook payload built from the
merge request reflected its state before the service had run. By moving
the hook and integrations trigger to happen at the end the payload we
POST a pyaload that reflects the state of the MR after the service has
run.

Fixes
https://gitlab.com/gitlab-org/gitlab/-/issues/327452 where the
merge_status property was still `"preparing"`.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/143318#note_1752373882

Changelog: fixed