Replace UI usages of shared runner with instance runner
This change aims to reduce the inconsistency between the terminology of
"shared" and "instance" runner in our UI.

It is meant as a first step to making the UI more consistent and it is
not extensive, as there may be usages of "shared runner" left. It is
also not meant to cover internal changes in the code.

Changelog: changed