Coverage artifacts should be optional for as-if-foss
We might not be running all RSpec jobs therefore some artifact collectors
won't be available. This aligns with other artifact collectors.

However, `needs.optional` is not available when cross project pipeline
is involved. To work around this, always let the collectors run for the
as-if-foss pipeline. All artifacts are optional so this should be fine.