Increase MAX_PARALLEL_REMOTE_REQUESTS to 4
After rolling out the feature flag ci_parallel_remote_includes, we saw
that there was not problem with parallelizing the HTTP requests.
However, we did not see so much performance gain from it.

Remote HTTP includes may not affect the overall performance of
CI includes so we did not see a significant performance improvement.
However, it should be safe and better to increase the number of parallel
HTTP requests per pipeline.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/351250