Detect usage of pinned nav item and handle section expanded state
If a pinned nav item was used, the Pinned section will show as
expanded and the item's original section shows as collapsed.

If an item was clicked in its original section, this section shows as
expanded. As before, the pinned section still can show as expanded
because we remember its state seperately in a cookie.

Changelog: changed