Remove remaining `@include` directives from SCSS files
This replaces some more `@include` rules with their actual CSS
definitions. This is done in preparation for moving to Tailwind CSS,
which won't let us include utilities the same way due to our inability
to support the `@apply` directive.

The `@include` approach was already questioned in the past as it doesn't
have much value over explicit style definitions and just adds an
unnecessary abstraction layer.