docs: Make the leading slash requirement for filepath clear
Filepath _currently_ requires a leading slash, most likely because
the trailing slash after `downloads` is missing somewhere.

While we wait for this bug to be fixed (if ever, due to backwards
compatibility), make this extremely clear in the documentation.

Technically, this is not just a documentation clarification, but
also a documentation bug, as the link is not `...downloads/:filepath`
which would indicate the result would resolve to `...downloads//<path>`.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>