Updated saved replies services
Updated the saved replies services to allow the service to accept groups as well
as users.
The service will check if the model supports saved replies. For users this
will always be `true`, for groups we need to do a license check.
This should also allow in the future for projects to support saved replies.

https://gitlab.com/gitlab-org/gitlab/-/issues/330098