gitaly: Rename `UnknownRefs` error to `ReferenceNotFoundError`
To differentiate Gitatly RPC errors, structured errors with embedded
error details are returned to clients. When a reference is not found in
a repository, the `ReferenceNotFoundError` detail is used. GitLab Rails
currently maps this error detail to the `UnknownRefs` error.

Rename the error to `ReferenceNotFoundError` to better correlate with
the RPC error detail.