Fix two racing conditions related to JavaScript
When JavaScript is enabled, the password field is replaced by a Vue
component (has a eye icon button to toggle the visibility). Sometimes
test don't pass because the password is added to the password field and
then replaced by the Vue component.

Fixed another issue for WebAuthn, where clicking `Sign in via 2FA code`
doesn't enables to enter the verification code because the JQuery
functionality has not yet been initialized.

Fixed both issue by waiting to find elements or selectors that only
happens when the JavaScript is ready.