Fix gitlab-housekeeper default reviewers to fix finalize BG migration
This fixes the keep `OverdueFinalizeBackgroundMigration` which was
previously getting an error because `reviewers` is nil. Instead of
handling `nil` all over the place we can just set our arrays to default
empty values if they are optional. This also simplifies callers as they
can just append to arrays if it's cleaner.