Add workers to sync `sbom_occurrences.traversal_ids`
We are denormalizing `namespaces.traversal_ids` to sbom_occurrences in
order to remove an expesive subquery from dependency list queries. This
change allows us to sync that column when a Project or Group is
transfered. Event subscriptions will be created in a different MR to
avoid issues with canary deployments.