Add not null constraint to noteable_type
noteable_type is already validated for prescene
in the application code and this commit
adds the corresponding DB constraint to
prevent invalid notes records.

It also contains a migration to create
an index asynchronously that will be used to
locate, backup and remove the notes records
with null noteable_type for .com.

Changelog: other