Move ProtectedBranch::CacheService#refresh logic
Prior to this we implemented logic to refresh the cache correctly in the
BaseService by looping over a Group's Projects when we refresh the
Group's cache. This meant that it was possible to bypass this logic when
calling the CacheService directly.

This change moves the logic for refreshing the cache for a group's
projects into the cache service to ensure they are always refreshed when
a group is refreshed.