Update specs for hash interpolation
Because we are now planning to build out hash interpolation instead of
moving to text interpolation, we need to update the specs to disable the
text interpolation feature flag and accurately test hash interpolation.

This commit also tries to preserve some of the minor improvements to the
interpolation classes that were introduced along with the text
interpolation classes.

https://gitlab.com/gitlab-org/ci-cd/pipeline-authoring/-/issues/127