repository: Disable file history rename detection
When viewing the commit history of a file, the `FindCommits` Gitaly RPC
is invoked with the `follow` option enabled. This instructs the
underlying `git-log(1)` operation to follow file renames when fetching
commit history. This rename detection can be rather expensive for
repositories with deep commit history and result in long file commit
history load times.

Introduce the `remove_file_commit_history_following` feature flag which,
when enabled, prevents rename detection from being performed when
finding commits for a single file.