Fix flaky test `lazy_user_notes_count_aggregate_spec.rb`
This commit modifies the `lazy_user_notes_count_aggregate_spec.rb`
spec to de-duplicate the expected array of `ids`.

This is because, as currently written, if `FactoryBot` assigns an id
of 20 (or 10 or 30) to the `vulnerability`, the spec will fail, making
this a flaky spec.

You can see my [comment][0] on the issue for a more detailed breakdown
of the flakiness

[0]:https://gitlab.com/gitlab-org/gitlab/-/issues/441189#note_1765583669

resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/441189