Prepare rails Web tests for gitaly stopping using infor/attribute
In Epic https://gitlab.com/groups/gitlab-org/-/epics/9006,
gitaly is removing its reliance on writing out
info/gitattributes. A key step of that is to stop
using ApplyGitattributes() (make it no-op) see MR
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/6573.

ApplyGitattributes() is the RPC call behind copy_gitattributes,
so this MR aims to prepare rails tests for
stopping using ApplyGitattributes().

The purpose of this MR has two folds:

1. with this changes, the rail tests should pass with current gitaly
2. once MR https://gitlab.com/gitlab-org/gitaly/-/merge_requests/6573
is in, the tests should pass with updated gitaly