Migrate some commits styles to a page-specific CSS bundle
This migrates commits lists styles to a page-specific CSS bundle. Those
styles are used in projects' commits list as well as in merge requests
commits.

The migration is done behind a feature flag. In the next iteration,
we'll move this first part out of the feature flag guard, and we'll
migrate remaining styles to another page-specific bundle as what remains
is specific to a single commit's details.