Static webhook payloads for snippet and label
This defines static webhook payloads for snippets and labels. This means
as attributes are added to these models we do not automatically
include them in the webhook payload until we redefine their `#hook_attr`
methods to do so.

This is a best practice to avoid sending data that we do not expect in
future.

The changes are behind feature flags.

https://gitlab.com/gitlab-org/gitlab/-/issues/440384