Show private group invited members to shared group members
We're doing this to enable shared group members to see who all can
access the group which will enable better collaboration among all the
group members.

Before we enable the added feature flag we will make changes to hide
the source of membership if the current user is not part of the
private invited group.