Show private invited group members to shared project members
We're doing this to enable shared project members to see who all can
access the project which will enable better collaboration among all the
project members.

Before we enable the feature flag we will make changes to hide the
source of membership of the current user is not part of the private
invited group.

We have done the same for group in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/143651.