Make Runner Fleet backend reviews optional
- We already have mandatory reviews from the verify team
for the most of our changes.
Most of the time Verify reviewers are qualified enough to review our code.
- Adding additional mandatory CODEOWNERS section creates
additional friction.
- We only have 2 Rails developers on the runner team,
so it's hard to find reviewers when someone is on PTO.

See https://gitlab.com/gitlab-org/gitlab/-/merge_requests/143214#note_1766708808
as an example.

So I suggest we make Runner team reviews optional.

Also, refactored this to use CODEOWNERS section.