Make sure new CSS compiler respects HMR setting
We run tests in CI with `vite` (`e2e:test-on-gdk`). When these run in
CI, we do not need to watch our CSS assets, because the CSS is not going
to be changed during the run of the pipeline. Therefore we just make the
CSS compiler respect the hot module reloading setting.