Update spec logic to avoid validation
This commit is part of a changeset that prevents release.sha column from
being updated. This test failed as a result of the change. This commit
updates the test logic so avoid model validation. This seems safe in
this situation since this spec is not designed to test update logic, but
is designed to test the return value when a SHA is nil, so changing the
update logic here does not change the nature of the spec.