Adds property to agent_version_id to chat aiAction
When agent_version_id is present, the system prompt in chat will be
replaced from the default duo chat to the prompt defined by the user on
the provided agent. Using this API requires users to be able to use chat
in itself, in addition to being able to use ai_agents feature.