Remove unused code
Originally, `DiffTree` was introduced to handle diffs for the initial
commit. Gitaly didn't return a diff of the original commit when it was
provided to FindChangedPaths RPC.

To solve that I implemented a fix:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/123950. It allows
to provide a `DiffTree` instance to the RPC and use a tree interface to
fetch changes even for the initial commit.

Later, a better option was discovered:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/143216.

As a result, some old code is not needed anymore.

Changelog: other